requirements.txt
asgiref==3.5.1
astroid==2.9.3
async-generator==1.10
attrs==21.4.0
awscli==1.22.57
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
black==21.12b0
bleach==5.0.0
boto==2.49.0
botocore==1.24.2
certifi==2021.10.8
cffi==1.15.0
click==7.1.2
colorama==0.4.3
cryptography==36.0.1
datadispatch==1.0.0
distlib==0.3.3
Django==4.0.4
djangorestframework==3.13.1
docutils==0.15.2
drf-api-logger==1.1.10
fabric==2.6.0
filelock==3.3.2
graphviz==0.19.1
h11==0.12.0
idna==3.3
invoke==1.6.0
isort==5.10.1
jmespath==0.10.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.4
outcome==1.1.0
pandas==1.3.4
paramiko==2.9.2
pathlib2==2.3.7.post1
pathspec==0.9.0
platformdirs==2.4.0
prettytable==3.1.1
pyasn1==0.4.8
pycparser==2.21
pylint==2.12.2
PyNaCl==1.5.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tasks==0.1.0
tinydb==3.15.1
toml==0.10.2
tomli==1.2.3
trio==0.19.0
trio-websocket==0.9.2
typer==0.4.0
typing_extensions==4.0.1
tzdata==2022.1
urllib3==1.26.8
virtualenv==20.10.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
wsproto==1.0.0
